Bug: Fixed issue with play/pause button not changing. Worked on song length slider.

+++ app/src/main/java/com/example/strawberry_app/data/dao/PlaylistSongDao.kt
@@ -13,7 +13,7 @@ data class SongWithPosition(
     val artist: String,
     val album: String,
     val length: Long,
-    val position: Int
+    val position: Long
 )
 
 @Dao
